+++ src/mellotron.ts
@@ -2,7 +2,7 @@ import { ChannelOptions } from "./player/channel";
 import { DefaultPlayer } from "./player/default-player";
 import {
   createEmptyRegionGroup,
-  findFirstSampleInLayer,
+  findFirstSampleInRegions,
   spreadRegions,
 } from "./player/layers";
 import {
@@ -76,7 +76,7 @@ export type MellotronOptions = Partial<
 export class Mellotron implements InternalPlayer {
   private readonly config: MellotronConfig;
   private readonly player: DefaultPlayer;
-  private readonly layer: RegionGroup;
+  private readonly group: RegionGroup;
   readonly load: Promise<this>;
 
   public constructor(
@@ -85,12 +85,12 @@ export class Mellotron implements InternalPlayer {
   ) {
     this.config = getMellotronConfig(options);
     this.player = new DefaultPlayer(context, options);
-    this.layer = createEmptyRegionGroup(options);
+    this.group = createEmptyRegionGroup(options);
 
     const loader = loadMellotronInstrument(
       this.config.instrument,
       this.player.buffers,
-      this.layer
+      this.group
     );
     this.load = loader(context, this.config.storage).then(() => this);
   }
@@ -104,8 +104,8 @@ export class Mellotron implements InternalPlayer {
   }
 
   start(sample: SampleStart | string | number) {
-    const found = findFirstSampleInLayer(
-      this.layer,
+    const found = findFirstSampleInRegions(
+      this.group,
       typeof sample === "object" ? sample : { note: sample }
     );
 
@@ -132,7 +132,7 @@ function getMellotronConfig(
 function loadMellotronInstrument(
   instrument: string,
   buffers: AudioBuffers,
-  layer: RegionGroup
+  group: RegionGroup
 ) {
   let variation = INSTRUMENT_VARIATIONS[instrument];
   if (variation) instrument = variation[0];
@@ -154,7 +154,7 @@ function loadMellotronInstrument(
             loadAudioBuffer(context, sampleUrl, storage).then((audioBuffer) => {
               buffers[sampleName] = audioBuffer;
               const duration = audioBuffer?.duration ?? 0;
-              layer.regions.push({
+              group.regions.push({
                 sampleName: sampleName,
                 midiPitch: midi,
                 sample: {
@@ -168,7 +168,7 @@ function loadMellotronInstrument(
         )
       )
       .then(() => {
-        spreadRegions(layer.regions);
+        spreadRegions(group.regions);
       });
   };
 }