+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ## 0.8.1
 
 - Fix: soundfont.loop returns this
-- Improvement: Soundfont uses layer abstraction. Simplify code
+- Improvement: Soundfont uses group abstraction. Simplify code
 
 ## 0.8.0